fix(calm-hub): address review findings on the cache service slice
- Collapse CalmCacheService/CaffeineCacheService into a single concrete
  CalmCacheService bean. calm-hub puts an interface in front of a
  service only where multiple backends are selected at runtime (see
  store/ and its Mongo/Nitrite producers); a cache with one
  implementation and no near-term second one doesn't fit that pattern.
- Add getList(key, elementType), so the one known consumer
  (GitHubVersionService, landing in slice 5) doesn't need
  get(key, List.class) plus an unchecked cast to use a typed list.
- Document the class contract: null values are ignored on put, a type
  mismatch on get/getList returns empty rather than throwing, and TTL
  is not refreshed on read.
- Reject a null ttl in put() with a clear NPE at the call site instead
  of failing later inside Caffeine's Expiry callback.
- Make maximumSize configurable via calm.cache.max-size (default
  10000, constructor-injected) instead of hardcoded, following the
  module's @ConfigProperty convention.
- Make the Caffeine Ticker injectable via a package-private
  constructor, matching the LongSupplier pattern already used by
  SchemaMigrationInProgressFilter, and use it to make TTL expiry
  tests deterministic instead of Thread.sleep.
- Rewrite the concurrent-access test to assert on final cache state
  via the submitted Futures, instead of only checking a
  CountDownLatch that a finally-block would trip even if every task
  had thrown.
