+++ calm-hub/src/main/resources/application.properties
@@ -17,6 +17,11 @@ calm.standalone.seed-demo-data=false
 # Read-only mode: when true, opens Nitrite with readOnly(true) and rejects mutating HTTP verbs
 calm.readonly=false
 
+# Maximum number of entries CalmCacheService (org.finos.calm.cache) holds at
+# once, evicted LRU-first once exceeded. TTL is set per entry at the call
+# site, not here.
+calm.cache.max-size=10000
+
 # Audit logging: independently toggle persistence to the auditLogs store and
 # emission of a structured log line under the org.finos.calm.audit category.
 # Both default to true. sourceIp capture defaults to false and has its own
